--- CheckInTerminal/bin/flightCheckIn.bash
@@ -1,3 +1,0 @@
-javac ../src/*.java -d .
-java CheckInTerminal.FlightCheckIn > output.txt
-
