+++ CheckInTerminal/src/part2/CheckinCounter.java
@@ -1,0 +1,4 @@
+package part2;
+public class CheckinCounter {
+    
+}
