+++ CheckInTerminal/src/part2/PassengerQueue.java
@@ -1,0 +1,4 @@
+package part2;
+public class PassengerQueue {
+
+}
